Add sound

+++ main scenes/Audio.cs
@@ -3,13 +3,13 @@ using System;
 
 public class Audio : Node
 {
-	public void PlaySound(string path, int minPitch = 1, int maxPitch = 1)
+	public void PlaySound(string path, float minPitch = 1, float maxPitch = 1)
 	{
 		AudioStreamPlayer audio = new AudioStreamPlayer();
 		audio.Stream = GD.Load<AudioStream>("res://assets/Music/" + path);
 		AddChild(audio);
         Random random = new Random();
-        audio.PitchScale = random.Next(minPitch * 100 , maxPitch * 100)/100.0f;
+        audio.PitchScale = random.Next((int)(minPitch * 100) , (int)(maxPitch * 100))/100.0f;
 		audio.Play();
 	}
 }
